api/requirements.txt
ipykernel
python-dotenv
pandas
pydantic
fastapi
uvicorn
openpyxl
python-multipart
identity>=0.5.1,<0.6
itsdangerous
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0